overwrite lemi424.py with lemi424_new.py
